refactor(football): extract TeamCaptionBuilder and LineupBuilder from FootballTeamCommand
